Clean up API

+++ lib/swift-extensions/SwiftExtensions/RandomNumber.swift
@@ -12,13 +12,13 @@ import Darwin
 /// By default a `RandomNumber` represents a value between 0 and 9, inclusive.
 public struct RandomNumber {
 
-  public typealias Value = Int32
+  public typealias Value = UInt32
 
   // MARK: Lifecycle
 
   /// - param start: The start of the inclusive range of allowed values.
   /// - param start: The number of allowed values.
-  public init(start: Int = 0, count: Int32 = 10) throws {
+  public init(start: Int = 0, count: UInt32 = 10) throws {
     guard start >= 0 else { throw RandomNumberError.notImplemented }
     self.start = start
     self.count = count
@@ -33,13 +33,13 @@ public struct RandomNumber {
 
   public lazy var value: Value = {
     // Eventually I will want to shift this value based on `start`.
-    return Int32(arc4random_uniform(UInt32(count)))
+    return arc4random_uniform(count)
   }()
 
   // MARK: Private
 
   private let start: Int
-  private let count: Int32
+  private let count: UInt32
 }
 
 // MARK: - RandomNumberError
